Extract all the AR setup from integration_spec to a new support file
